Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/013-expected.txt


FAIL Link with onclick navigation to javascript url with delayed document.write and href navigation  assert_array_equals: expected property 0 to be "after script" but got "javascript" (expected array ["after script", "javascript"] got ["javascript", "after script"])
 Test
